Move factory method create to Location class
